converted everything to ts
