+++ src/Constants.ts
@@ -8,6 +8,9 @@ import { IMap } from "./types";
 export const CONSTANTS: IMap<any> = {
     Version:                "0.52.2",
 
+    // Speed (in ms) at which the main loop is updated
+    _idleSpeed: 200,
+
     /** Max level for any skill, assuming no multipliers. Determined by max numerical value in javascript for experience
      * and the skill level formula in Player.js. Note that all this means it that when experience hits MAX_INT, then
      * the player will have this level assuming no multipliers. Multipliers can cause skills to go above this.
