+++ src/Gang/data/Constants.ts
@@ -10,7 +10,7 @@ export const GangConstants: {
     MaximumGangMembers: 30,
     CyclesPerTerritoryAndPowerUpdate: 100,
      // Portion of upgrade multiplier that is kept after ascending
-    AscensionMultiplierRatio: 15,
+    AscensionMultiplierRatio: .15,
     // Names of possible Gangs
     Names: [
         "Slum Snakes",
