+++ src/Prestige.js
@@ -14,6 +14,7 @@ import { Engine } from "./engine";
 import { Faction } from "./Faction/Faction";
 import { Factions, initFactions } from "./Faction/Factions";
 import { joinFaction } from "./Faction/FactionHelpers";
+import { clearGangUI } from "./Gang/Helpers";
 import { updateHashManagerCapacity } from "./Hacknet/HacknetHelpers";
 import { initMessages } from "./Message/MessageHelpers";
 import { prestigeWorkerScripts } from "./NetscriptWorker";
@@ -333,7 +334,7 @@ function prestigeSourceFile(flume) {
         deleteStockMarket();
     }
 
-    if (Player.inGang()) { Player.gang.clearUI(); }
+    if (Player.inGang()) clearGangUI();
     Player.gang = null;
     Player.corporation = null; resetIndustryResearchTrees();
     Player.bladeburner = null;
